refactor: pg pass-through socket utils
